Move WeakFrechet to Distances.h

Squashed commits:

[ac52803] Move WeakFrechet to Distances.h
